+++ parlwin/tests/Command/SyncCommandTest.php
@@ -265,6 +265,10 @@ class SyncCommandTest extends TestCase {
             public function setAppValue(string $app, string $key, string $value): void {
                 $this->values[$key] = $value;
             }
+
+            public function getSystemValue(string $key, $default = '') {
+                return $this->values[$key] ?? $default;
+            }
         };
 
         $realtimeEvents = [];
@@ -355,6 +359,10 @@ class SyncCommandTest extends TestCase {
 
             public function setAppValue(string $app, string $key, string $value): void {
                 $this->values[$key] = $value;
+            }
+
+            public function getSystemValue(string $key, $default = '') {
+                return $this->values[$key] ?? $default;
             }
         };